Env installation back to old method
After repeated attemts to get pyproject.toml to work, I gave up. The main problem is that the project folder doesn't follow GitHub's default folder structure, here pyproject.toml would be in the same folder as .git. It's stupid that this can't be fixed, but it is what it is. ON THE BRIGHT SIGHT pip install -e . --use-pep517 removes the deprecation warning. So this method will be valid with new conda, which was the main point of this.